requirements.txt
gspread==5.7.2
google-auth==2.22.0
requests==2.31.0
python-dotenv==1.0.0
